+++ Chapter16/Chapter16-1-2/SequentialFileProcessor.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.IO;
-
-namespace Chapter16_1_2 {
+﻿namespace Chapter16_1_2 {
     /// <summary>
     /// 非並列処理を行うロジックをまとめたクラス
     /// </summary>
@@ -13,31 +10,7 @@ namespace Chapter16_1_2 {
         /// <param name="vAllCsFilesPaths">C#ソースファイルのファイルパス一覧</param>
         public void ProcessFiles(string[] vAllCsFilesPaths) {
             foreach (string wCsFile in vAllCsFilesPaths) {
-                SearchCsFiles(wCsFile);
-            }
-        }
-        /// <summary>
-        /// 「async」と「await」の両方を含むファイルを検索するメソッド
-        /// </summary>
-        /// <param name="vCsFile">C#のソースファイル</param>
-        static void SearchCsFiles(string vCsFile) {
-            try {
-                var wCsFileContents = File.ReadAllText(vCsFile);
-                if (wCsFileContents.Contains("async") && wCsFileContents.Contains("await")) {
-                    Console.WriteLine(Path.GetFullPath(vCsFile));
-                }
-            }
-            catch (UnauthorizedAccessException wEx) {
-                Console.WriteLine($"ファイルへのアクセスが拒否されました：{wEx.Message}");
-            }
-            catch (PathTooLongException wEx) {
-                Console.WriteLine($"ファイルパスが長すぎます：{wEx.Message}");
-            }
-            catch (OutOfMemoryException wEx) {
-                Console.WriteLine($"メモリが不足しています：{wEx.Message}");
-            }
-            catch (Exception wEx) {
-                Console.WriteLine($"エラーが発生しました：{wEx.Message}");
+                CsFileSearcher.SearchCsFiles(wCsFile);
             }
         }
     }
